[NOTIFICATIONS] Add Notifications API's

+++ platform/microservices/trust-certification-system-certificate-events-processor/src/main/java/com/dreamsoftware/tcs/service/ICertificateGenerator.java
@@ -17,6 +17,6 @@ public interface ICertificateGenerator {
      * @return
      * @throws Exception
      */
-    File generate(String caName, String studentName, String courseName, Float qualification) throws Exception;
+    File generate(String caName, String studentName, String courseName, Long qualification) throws Exception;
 
 }
